// frontend/src/app/api/model/bulkOperationResult.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BulkOperationResult { 
    /**
     * Number of ports successfully updated
     */
    success?: number;
    /**
     * Number of ports that failed
     */
    failed?: number;
    /**
     * Per-port error messages
     */
    errors?: Array<string>;
}