// frontend/src/app/api/model/pingRequest.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PingRequest { 
    /**
     * IPv4 address to ping from the switch
     */
    address: string;
    /**
     * Number of ICMP packets to send
     */
    count?: number;
    /**
     * Per-packet timeout in milliseconds
     */
    timeoutMs?: number;
    /**
     * ICMP packet size in bytes
     */
    size?: number;
}