// frontend/src/app/api/model/pingResult.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PingResult { 
    /**
     * Number of packets sent
     */
    sent?: number;
    /**
     * Number of packets received
     */
    received?: number;
    /**
     * Minimum round-trip time in milliseconds
     */
    minRtt?: number;
    /**
     * Average round-trip time in milliseconds
     */
    avgRtt?: number;
    /**
     * Maximum round-trip time in milliseconds
     */
    maxRtt?: number;
}